requirements.txt
django
pillow
django-allauth
django-crispy-forms
django-star-ratings
django-tiny-mce4